fixed minors problems in various files for first business test

+++ packages/notest-cluster-runner/src/service/cluster-runner.service.ts
@@ -198,6 +198,6 @@ export class ClusterRunnerService {
   }
 
   private removeVideo(videoPath: string) {
-    fs.rmSync(videoPath, { recursive: true, force: true });
+    if (fs.existsSync(videoPath)) fs.rmSync(videoPath, { recursive: true, force: true });
   }
 }
